server/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
bootstrap4==0.1.0
djangorestframework-simplejwt==5.2.2
django-extensions==3.1.3
